TSF-240 Renamer package name
