+++ packages/omsorgen-for-oms/i18n/nb_NO.json
