+++ packages/omsorgen-for-oms/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@k9-sak-web/omsorgspenger-omsorgen-for",
+  "name": "@k9-sak-web/omsorgen-for-oms",
   "version": "1.0.0",
   "module": "index.tsx",
   "license": "MIT",
