+++ packages/omsorgen-for-oms/src/components/spørsmål.less
